+++ src/tag/SpeechPartTag.cpp
@@ -1,7 +1,7 @@
 #include "SpeechPartTag.h"
 namespace base {
-static const boost::bimap<uint128_t, std::string> SPEECH_PART_MAP =
-    boost::assign::list_of<boost::bimap<uint128_t, std::string>::relation>
+static const boost::bimap<uint64_t, std::string> SPEECH_PART_MAP =
+    boost::assign::list_of<boost::bimap<uint64_t, std::string>::relation>
     (0x00     ,     "UNKN")
     (0x01     ,     "NOUN")
     (0x02     ,     "ADJF")
@@ -21,24 +21,24 @@ static const boost::bimap<uint128_t, std::string> SPEECH_PART_MAP =
     (0x8000   ,     "PRCL")
     (0x10000  ,     "INTJ");
 
-const SpeechPartTag SpeechPartTag::UNKN(uint128_t(0x00   ));
-const SpeechPartTag SpeechPartTag::NOUN(uint128_t(0x01   ));
-const SpeechPartTag SpeechPartTag::ADJF(uint128_t(0x02   ));
-const SpeechPartTag SpeechPartTag::ADJS(uint128_t(0x04   ));
-const SpeechPartTag SpeechPartTag::COMP(uint128_t(0x08   ));
-const SpeechPartTag SpeechPartTag::VERB(uint128_t(0x10   ));
-const SpeechPartTag SpeechPartTag::INFN(uint128_t(0x20   ));
-const SpeechPartTag SpeechPartTag::PRTF(uint128_t(0x40   ));
-const SpeechPartTag SpeechPartTag::PRTS(uint128_t(0x80   ));
-const SpeechPartTag SpeechPartTag::GRND(uint128_t(0x100  ));
-const SpeechPartTag SpeechPartTag::NUMR(uint128_t(0x200  ));
-const SpeechPartTag SpeechPartTag::ADVB(uint128_t(0x400  ));
-const SpeechPartTag SpeechPartTag::NPRO(uint128_t(0x800  ));
-const SpeechPartTag SpeechPartTag::PRED(uint128_t(0x1000 ));
-const SpeechPartTag SpeechPartTag::PREP(uint128_t(0x2000 ));
-const SpeechPartTag SpeechPartTag::CONJ(uint128_t(0x4000 ));
-const SpeechPartTag SpeechPartTag::PRCL(uint128_t(0x8000 ));
-const SpeechPartTag SpeechPartTag::INTJ(uint128_t(0x10000));
+const SpeechPartTag SpeechPartTag::UNKN(uint64_t(0x00   ));
+const SpeechPartTag SpeechPartTag::NOUN(uint64_t(0x01   ));
+const SpeechPartTag SpeechPartTag::ADJF(uint64_t(0x02   ));
+const SpeechPartTag SpeechPartTag::ADJS(uint64_t(0x04   ));
+const SpeechPartTag SpeechPartTag::COMP(uint64_t(0x08   ));
+const SpeechPartTag SpeechPartTag::VERB(uint64_t(0x10   ));
+const SpeechPartTag SpeechPartTag::INFN(uint64_t(0x20   ));
+const SpeechPartTag SpeechPartTag::PRTF(uint64_t(0x40   ));
+const SpeechPartTag SpeechPartTag::PRTS(uint64_t(0x80   ));
+const SpeechPartTag SpeechPartTag::GRND(uint64_t(0x100  ));
+const SpeechPartTag SpeechPartTag::NUMR(uint64_t(0x200  ));
+const SpeechPartTag SpeechPartTag::ADVB(uint64_t(0x400  ));
+const SpeechPartTag SpeechPartTag::NPRO(uint64_t(0x800  ));
+const SpeechPartTag SpeechPartTag::PRED(uint64_t(0x1000 ));
+const SpeechPartTag SpeechPartTag::PREP(uint64_t(0x2000 ));
+const SpeechPartTag SpeechPartTag::CONJ(uint64_t(0x4000 ));
+const SpeechPartTag SpeechPartTag::PRCL(uint64_t(0x8000 ));
+const SpeechPartTag SpeechPartTag::INTJ(uint64_t(0x10000));
 
 
 const std::vector<SpeechPartTag> SpeechPartTag::inner_runner = {
@@ -62,10 +62,10 @@ const std::vector<SpeechPartTag> SpeechPartTag::inner_runner = {
     INTJ,
 };
 
-SpeechPartTag::SpeechPartTag(uint128_t val): ITag(val, &SPEECH_PART_MAP) {}
+SpeechPartTag::SpeechPartTag(uint64_t val): ITag(val, &SPEECH_PART_MAP) {}
 
 SpeechPartTag::SpeechPartTag(const std::string &val): ITag(val, &SPEECH_PART_MAP) {}
 
-SpeechPartTag::SpeechPartTag(): ITag((uint128_t)0, &SPEECH_PART_MAP) {}
+SpeechPartTag::SpeechPartTag(): ITag((uint64_t)0, &SPEECH_PART_MAP) {}
 }
